avoid 0

+++ Editor/EventLogger.cs
@@ -180,7 +180,7 @@ namespace StatsigUnity
         {
             while (true)
             {
-                yield return new WaitForSeconds(delay / 1000);
+                yield return new WaitForSeconds((float) (delay / 1000f));
                 FlushEvents(false);
             }
         }
